Address PR review feedback for summarize
